second day of api route management

+++ server/app/Http/Controllers/Api/SchedulesController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\ScheduleResource;
 use App\Models\DoctorProfile;
 use App\Models\Room;
 use App\Models\Schedule;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
@@ -22,8 +24,8 @@ class SchedulesController extends Controller
 
     public function index()
     {
-        $schedule = Schedule::all();
-        return response()->json([ 'status' => 200,'data' => $schedule], 200);
+        $schedules = ScheduleResource::collection(Schedule::all());
+        return response()->json([ 'status' => 200,'data' => $schedules], 200);
     }
 
     // // if needed to create 
@@ -40,11 +42,22 @@ class SchedulesController extends Controller
     //     ]);
     // }
 
+    public function show(Schedule $schedule){
+
+            return response()->json([
+                'status' => 200,
+                'data' => new ScheduleResource($schedule),
+            ], 200);
+        
+    }
+
     public function store(Request $request)
     {
+        Gate::authorize('create', Schedule::class);
+
         $validator = Validator::make($request->all(), [
-            'room_id'=>"nullable|exists:rooms,id",
-            'doctor_id'=>"nullable|exists:doctor_profiles,id",
+            'room_id'=>"required|exists:rooms,id",
+            'doctor_id'=>"required|exists:doctor_profiles,id",
             'day'=> 'required',
             'period'=> 'required',
             'booking_limit'=> 'required'
@@ -67,7 +80,8 @@ class SchedulesController extends Controller
 
             return response()->json([
                 'status' => 200,
-                'message' => 'Room created successfully'
+                'data' => new ScheduleResource($schedule),
+                'message' => 'schedule created successfully'
             ], 200);
 
         } catch (\Exception $e) {
@@ -91,11 +105,12 @@ class SchedulesController extends Controller
     //     ]);
     // }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, Schedule $schedule)
     {
+        Gate::authorize('update', $schedule);
         $validator = Validator::make($request->all(), [
-            'room_id'=>"nullable|exists:rooms,id",
-            'doctor_id'=>"nullable|exists:doctor_profiles,id",
+            'room_id'=>"exists:rooms,id",
+            'doctor_id'=>"  exists:doctor_profiles,id",
             'day'=> 'required',
             'period'=> 'required',
             'booking_limit'=> 'required'
@@ -107,7 +122,6 @@ class SchedulesController extends Controller
 
         try {
 
-            $schedule = Schedule::findOrFail($id);
             $schedule->room_id = $request->room_id;
             $schedule->doctor_id = $request->doctor_id;
             $schedule->day = $request->day;
@@ -118,6 +132,7 @@ class SchedulesController extends Controller
 
             return response()->json([
                 'status' => 200,
+                'data' => new ScheduleResource($schedule),
                 'message' => 'Schedule updated successfully'
             ], 200);
 
@@ -126,15 +141,14 @@ class SchedulesController extends Controller
         }
     }
 
-    public function destroy(string $id)
+    public function destroy(Schedule $schedule)
     {
-        try {
-            $schedule = Schedule::findOrFail($id);
+        Gate::authorize('delete', $schedule);
 
+        try {
             if (!$schedule) {
                 return $this->errorResponse('Schedule not found', null, 404);
             }
-
             $schedule->delete();
 
             return response()->json([
